+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * **Add**: Telemetry events for all operation calls.
   See the [telemetry guide](guides/telemetry.md) for more information.
+* **Add**: OpenTelemetry plugin for easy integration with existing telemetry events.
 
 ### 0.0.13 (2023-09-12)